Move Equal, EqualIPv6 and Patchable into forwardingrules pkg
Mixed protocol feature requires those functions. To prevent package dependency cycles these methods are moved to pkg/forwardingrules.
